requirements.txt
fastapi==0.100.1
pydantic-settings==2.0.2

alembic==1.11.1
sqlalchemy==2.0.19
asyncpg==0.28.0

uvicorn[standard]==0.23.2

flake8==6.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
